Remove symbolic in task-spec

+++ lib/task-spec/test/src/task-spec/dynamic_graph/shard_expansion.cc
@@ -103,7 +103,7 @@ TEST_SUITE(FF_TEST_SUITE) {
       };
     };
 
-    auto mk_value = [](int src_node_id, 
+    auto mk_value = [](size_t src_node_id, 
                        TensorSlotName src_slot_name, 
                        std::optional<ParallelTensorSpaceCoordinate> const &shard_coord) 
       -> DynamicValueAttrs
